+++ wiki/CHANGES.md
@@ -1,5 +1,14 @@
 # Changes
 
+## 2026-01-29
+
+- [2.2.0] Change to hook ```bajo:afterBoot```
+
+## 2026-01-27
+
+- [2.1.1] Remove dependency to ```masohi``` so you can directly tap mqtt events without ```masohi```
+- [2.1.1] Change ```instance``` to ```client``` to match other the entire bajo ecosystem convention
+
 ## 2026-01-08
 
 - [2.1.0] Ported to ```bajo@2.2.x``` & ```dobo@2.2.x``` specs